+++ packages/eslint-plugin-mobx/__tests__/TODO-no-anonymous-observer.js
@@ -1,0 +1,15 @@
+import { RuleTester } from "eslint";
+
+import rule from "../src/no-anonymous-observer.js";
+
+const tester = new RuleTester({
+  parser: require.resolve('@typescript-eslint/parser'),
+  parserOptions: {}
+});
+
+tester.run("no-anonymous-observer", rule, {
+  valid: [
+    { code: `` },
+  ],
+  invalid: [],
+});