+++ packages/mobx/__tests__/v5/base/observables.js
@@ -15,8 +15,9 @@ const {
     isObservableProp
 } = mobx
 const utils = require("../../v5/utils/test-utils")
+const { MAX_SPLICE_SIZE } = require("../../../src/internal")
 
-const voidObserver = function () {}
+const voidObserver = function () { }
 
 function buffer() {
     const b = []
@@ -2099,7 +2100,7 @@ test("extendObservable should not accept complex objects as second argument", ()
 })
 
 test("observable ignores class instances #2579", () => {
-    class C {}
+    class C { }
     const c = new C()
     expect(observable(c)).toBe(c)
 })
@@ -2120,9 +2121,9 @@ test("configure({ safeDescriptors: false })", () => {
 
     class Clazz {
         observable = 0
-        action() {}
-        get computed() {}
-        *flow() {}
+        action() { }
+        get computed() { }
+        *flow() { }
         constructor() {
             mobx.makeObservable(this, {
                 observable: mobx.observable,
@@ -2138,9 +2139,9 @@ test("configure({ safeDescriptors: false })", () => {
 
     const plain = mobx.observable({
         observable: 0,
-        action() {},
-        get computed() {},
-        *flow() {}
+        action() { },
+        get computed() { },
+        *flow() { }
     })
 
     checkDescriptors(plain)
@@ -2189,4 +2190,95 @@ test("options can be provided only once", () => {
         o.y = 0
         makeObservable(o, { y: observable }, {})
     }).toThrow(error)
+})
+
+test("ObservableArray.replace", () => {
+    // both lists are small
+    let ar = observable([1])
+    let del = ar.replace([2])
+    expect(ar.toJSON()).toEqual([2])
+    expect(del).toEqual([1])
+
+    // the replacement is large
+    ar = observable([1])
+    del = ar.replace(new Array(MAX_SPLICE_SIZE))
+    expect(ar.length).toEqual(MAX_SPLICE_SIZE)
+    expect(del).toEqual([1])
+
+    // the original is large
+    ar = observable(new Array(MAX_SPLICE_SIZE))
+    del = ar.replace([2])
+    expect(ar).toEqual([2])
+    expect(del.length).toEqual(MAX_SPLICE_SIZE)
+
+    // both are large; original larger than replacement
+    ar = observable(new Array(MAX_SPLICE_SIZE + 1))
+    del = ar.replace(new Array(MAX_SPLICE_SIZE))
+    expect(ar.length).toEqual(MAX_SPLICE_SIZE)
+    expect(del.length).toEqual(MAX_SPLICE_SIZE + 1)
+
+    // both are large; replacement larger than original
+    ar = observable(new Array(MAX_SPLICE_SIZE))
+    del = ar.replace(new Array(MAX_SPLICE_SIZE + 1))
+    expect(ar.length).toEqual(MAX_SPLICE_SIZE + 1)
+    expect(del.length).toEqual(MAX_SPLICE_SIZE)
+})
+
+test("ObservableArray.splice", () => {
+    // Deleting 1 item from a large list
+    let ar = observable(new Array(MAX_SPLICE_SIZE + 1))
+    let del = ar.splice(1, 1)
+    expect(ar.length).toEqual(MAX_SPLICE_SIZE)
+    expect(del.length).toEqual(1)
+
+    // Deleting many items from a large list
+    ar = observable(new Array(MAX_SPLICE_SIZE + 2))
+    del = ar.splice(1, MAX_SPLICE_SIZE + 1)
+    expect(ar.length).toEqual(1)
+    expect(del.length).toEqual(MAX_SPLICE_SIZE + 1)
+
+    // Deleting 1 item from a large list and inserting many items
+    ar = observable(new Array(MAX_SPLICE_SIZE + 1))
+    del = ar.splice(1, 1, ...new Array(MAX_SPLICE_SIZE + 1))
+    expect(ar.length).toEqual(MAX_SPLICE_SIZE * 2 + 1)
+    expect(del.length).toEqual(1)
+
+    // Deleting many items from a large list and inserting many items
+    ar = observable(new Array(MAX_SPLICE_SIZE + 10))
+    del = ar.splice(1, MAX_SPLICE_SIZE + 1, ...new Array(MAX_SPLICE_SIZE + 1))
+    expect(ar.length).toEqual(MAX_SPLICE_SIZE + 10)
+    expect(del.length).toEqual(MAX_SPLICE_SIZE + 1)
+})
+
+describe("`requiresReaction` takes precedence over global `computedRequiresReaction`", () => {
+    let warnMsg = "[mobx] Computed value 'TestComputed' is being read outside a reactive context. Doing a full recompute.";
+    let consoleWarnSpy;
+    beforeEach(() => {
+        consoleWarnSpy = jest.spyOn(console, "warn").mockImplementation()
+    })
+    afterEach(() => {
+        consoleWarnSpy.mockRestore()
+        mobx._resetGlobalState();
+    })
+
+    test('`undefined`', () => {
+        mobx.configure({ computedRequiresReaction: true })
+        const c = mobx.computed(() => { }, { name: 'TestComputed' });
+        c.get();
+        expect(consoleWarnSpy).toHaveBeenLastCalledWith(warnMsg);
+    })
+
+    test('`true` over `false`', () => {
+        mobx.configure({ computedRequiresReaction: false })
+        const c = mobx.computed(() => { }, { name: 'TestComputed', requiresReaction: true });
+        c.get();
+        expect(consoleWarnSpy).toHaveBeenLastCalledWith(warnMsg);
+    })
+
+    test('`false` over `true`', () => {
+        mobx.configure({ computedRequiresReaction: true })
+        const c = mobx.computed(() => { }, { name: 'TestComputed', requiresReaction: false });
+        c.get();
+        expect(consoleWarnSpy).not.toHaveBeenCalled();
+    })
 })